requirements.txt
# Legion — zależności projektu
# Instalacja: pip install -r requirements.txt
#
# UWAGA: gi (PyGObject) musi być zainstalowane jako paczka systemowa:
#   sudo pacman -S python-gobject webkit2gtk-4.1 wl-clipboard tor
# Następnie symlink do venv:
#   ln -sf /usr/lib/python3.x/site-packages/gi .venv/lib/python3.x/site-packages/gi

# Kryptografia
pynacl>=1.5

# Sieć Tor
stem>=1.8

# Sieć asynchroniczna
websockets>=12.0

# API lokalne (dla GUI)
fastapi>=0.111
uvicorn>=0.29
httpx>=0.28        # testowanie API (httpx.AsyncClient)

# Baza danych
aiosqlite>=0.20

# Sanityzacja plików i obrazów (stripping metadanych EXIF/GPS/XMP)
pillow>=10.0

# GUI
pywebview>=5.0     # GTK backend — po instalacji wykonaj:
                   #   sudo pacman -S python-gobject webkit2gtk-4.1
                   #   PYVER=$(python3 -c "import sys; print(f'{sys.version_info.major}.{sys.version_info.minor}')")
                   #   ln -sf /usr/lib/python${PYVER}/site-packages/gi .venv/lib/python${PYVER}/site-packages/gi

# Dev / testy
pytest>=8.0
pytest-asyncio>=0.23
